refactor: remove the info plugin and add it as an option to list
